[refactor] : Section 컴포넌트를 활용한 Stt 컴포넌트 구성
