+++ src/containers/PbProfile.tsx
@@ -1,9 +1,9 @@
 import { Switch } from '@radix-ui/react-switch';
 import { useState, useRef } from 'react';
 import profileImage from '../assets/조경은PB.png';
+import Section from '../components/Section';
 import getOffice from '../lib/getOffice';
 import { type TPbProps } from '../lib/types';
-import Section from './Section';
 
 export default function PbProfile() {
   const [profile, setProfile] = useState<TPbProps>({
