+++ src/containers/WriteNoti.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from 'react';
-import { Button } from './ui/button';
+import { Button } from '../components/ui/button';
 
 type TCustomer = {
   id: number;
