+++ src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import copyright from '../assets/copyright.png';
 import logo from '../assets/diA_business.png';
-import Login from '../components/Login';
+import Login from '../containers/Login';
 
 export default function LoginPage() {
   return (
