+++ src/pages/MainPage.tsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from 'react';
 import logo from '../assets/diA.png';
-import ConsultationRequest from '../components/ConsultationRequest';
-import GuestList from '../components/GuestList';
-import PbCalendar from '../components/PbCalendar';
-import PbProfile from '../components/PbProfile';
-import ScheduledConsultationList from '../components/ScheduledConsultationList';
 import Section from '../components/Section';
+import ConsultationRequest from '../containers/ConsultationRequest';
+import GuestList from '../containers/GuestList';
+import PbCalendar from '../containers/PbCalendar';
+import PbProfile from '../containers/PbProfile';
+import ScheduledConsultationList from '../containers/ScheduledConsultationList';
 import { type TCustomersProps } from '../lib/types';
 import { type TRequestedConsultationsProps } from '../lib/types';
